+++ xrpl/utils/time_conversions.py
@@ -13,7 +13,7 @@ from xrpl.constants import XRPLException
 RIPPLE_EPOCH: Final[int] = 946684800
 """The "Ripple Epoch" of 2000-01-01T00:00:00 UTC"""
 
-MAX_XRPL_TIME: Final[int] = 2**32
+MAX_XRPL_TIME: Final[int] = 2 ** 32
 """The maximum time that can be expressed on the XRPL"""