+++ CHANGELOG.md
@@ -7,9 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[Unreleased]]
 ### Added:
+- Function to parse the final account balances from a transaction's metadata
+
+### Fixed:
+- Typing for factory classmethods on models
+
+### Fixed:
+- Use properly encoded transactions in `Sign`, `SignFor`, and `SignAndSubmit`
+
+## [1.6.0] - 2022-06-02
+### Added:
 - Support for dynamic fee calculation
-- Add function to parse account balances from a transaction's metadata
+- Function to parse account balances from a transaction's metadata
 - Better error handling for invalid client URL
+- Exported SubscribeBook
 
 ### Fixed
 - Resolve `txnNotFound` error with `send_reliable_submission` when waiting for a submitted malformed transaction